requirements.txt
streamlit>=1.28.0
snowflake-connector-python>=3.6.0
pandas>=2.0.0
python-dotenv>=1.0.0
tabulate>=0.9.0
